requirements.txt
mongoengine==0.27.0
pymongo==4.8.0
python-dotenv==1.0.1
telemongo==0.2.2
Telethon==1.34.0
